@@ -16,10 +16,9 @@ dependencies:
   get_it: ^7.6.7
   provider: ^6.1.2
   connectivity: ^3.0.3
+  webview_flutter: ^4.9.0
   dio: ^5.6.0
   http: ^0.13.1
-  #intl: ^0.19.0
-  #http: ^1.1.0
   freezed: ^2.5.7
   shared_preferences: ^2.3.1
   fancy_bottom_navigation: ^0.3.3
@@ -40,7 +39,6 @@ dependencies:
   firebase_core_platform_interface: ^5.2.0
   flutter_svg: any
   flutter_html: any
-  webview_flutter: ^3.0.1
   phone_number: ^2.1.0
   country_code_picker: ^3.0.0
   pin_code_fields: ^8.0.1
@@ -71,6 +69,7 @@ dependencies:
   pdf: ^3.11.1
   printing: ^5.13.1
   open_file: ^3.3.1
+
 
 dev_dependencies:
   flutter_test:
